backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
python-dotenv
pymongo
pyjwt
passlib
bcrypt
